pyproject.toml
[project]
name = "microphone-asr"
version = "0.1.0"
description = "Microphone ASR"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "webrtcvad-wheels>=2.0.14",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "requests>=2.32.4",
]

[tool.setuptools]
package-dir = { "" = "src" }

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
